+++ packages/frontend/src/components/MkRolePreview.vue
@@ -1,6 +1,11 @@
 <template>
 <MkA v-adaptive-bg :to="`/admin/roles/${role.id}`" class="_panel" :class="$style.root" tabindex="-1" :style="{ '--color': role.color }">
 	<div :class="$style.title">
+		<span :class="$style.icon">
+			<i v-if="role.isAdministrator" class="ti ti-crown" style="color: var(--accent);"></i>
+			<i v-else-if="role.isModerator" class="ti ti-shield" style="color: var(--accent);"></i>
+			<i v-else class="ti ti-user" style="opacity: 0.7;"></i>
+		</span>
 		<span :class="$style.name">{{ role.name }}</span>
 		<span v-if="role.target === 'manual'" :class="$style.users">{{ role.usersCount }} users</span>
 		<span v-else-if="role.target === 'conditional'" :class="$style.users">({{ i18n.ts._role.conditional }})</span>
@@ -31,6 +36,10 @@ const props = defineProps<{
 	display: flex;
 }
 
+.icon {
+	margin-right: 8px;
+}
+
 .name {
 	font-weight: bold;
 }
@@ -42,5 +51,6 @@ const props = defineProps<{
 
 .description {
 	opacity: 0.7;
+	font-size: 85%;
 }
 </style>